+++ src/hooks/useWidth.ts
@@ -1,0 +1,16 @@
+import React from 'react';
+
+export default function useWidth(): [React.RefObject<HTMLDivElement>, number | undefined] {
+  const ref = React.useRef<HTMLDivElement>(null);
+  const [width, setWidth] = React.useState(ref?.current?.clientWidth);
+  React.useEffect(() => {
+    function handleResize() {
+      setWidth(ref?.current?.clientWidth ?? width);
+    }
+    handleResize();
+    window.addEventListener('resize', handleResize);
+    return () => window.removeEventListener('resize', handleResize);
+  }, [width]);
+
+  return [ref, width];
+}
